Refactor iconography components and update imports for improved structure and locale handling
